Include toolchain dependencies in verilator_cc_library.
Previously the Verilator toolchain dependencies (e.g. verilator.o) were
not being included in the static library generated by
verilator_cc_library.  In most situations this wasn't an issue as the
missing dependencies were also included in the follow on binary linking.
However, in situations (with gcc) where the binary doesn't actually
reference the Verilator dependencies (as is the case with alu.cpp), the
library's toolchain depedencies aren't transitively included
(presumeably due to garbage collection) and fails to link.  This change
fixes #3.

In addition, the inclusion of Valu.h must be relative to the root of the
Bazel workspace.  This change is made to enable successful execution via
`bazel run //test/alu:alu_bin`.