Pass Verilator include directory to dependencies.
This change places the generated module's include directory on the
search path during compilation of dependencies.  The benefit of this is
two fold: first it allows the application code to avoid the odd
directory naming (V{module}.h/...) required for tree artifacts, second
it paves the way for multiple verilog_cc_libraries which contain the
same top module but are compiled differently (i.e. different params such
as -GW=16).  In this case the application simply includes V{module}.h
and the correct include directory is choosen based on the BUILD rules.