Change Verilator mirror to Github and update to 4.100.
This changes the Verilator mirror to Github from the original Verilator
FTP url (which is no longer advertised or accesible from the website).
A byproduct of this is all of the version hashes must change and will
require a rebuild for users.

Strangly the distribution provided on the Verilator website is slightly
different than the Github tags in that it seems to contain files
produced by the `autoconf`, `./configure`, and `make` steps.  Namely it
contains auto-generated files `src/config_rev.h` and
`include/verilated_config.h`.  The repository rule is updated to provide
the correct generation / substitutions for the necessary files.